Changeing fft model to a package that can (and is!) on pypi
