+++ r2sdf_fft/casper_r2sdf_fft_py/README.md
@@ -1,0 +1,2 @@
+# CASPER R2SDF FFT Model
+This is a fixed point model of the CASPERHDL FFT block
