+++ r2sdf_fft/casper_r2sdf_fft_py/pyproject.toml
@@ -1,0 +1,26 @@
+[project]
+name = "casper_r2sdf_fft"
+version = "0.0.3"
+authors = [
+  { name="Matthew Schiller", email="mschille@nrao.edu" },
+]
+description = "A model for the CASPER FFT"
+readme = "README.md"
+requires-python = ">=3.8"
+classifiers = [
+    "Programming Language :: Python :: 3",
+    "License :: OSI Approved :: Apache Software License",
+    "Operating System :: OS Independent",
+]
+
+[project.urls]
+Homepage = "https://github.com/talonmyburgh/casper_dspdevelt"
+Issues = "https://github.com/talonmyburgh/casper_dspdevel/issues"
+
+[build-system]
+requires = ["hatchling","numpy","pathlib"]
+build-backend = "hatchling.build"
+
+[tool.hatch.build.targets.wheel]
+packages = ["src/casper_r2sdf_fft"]
+
